1st Iteration Complete
The app can create tasks, delete tasks, complete tasks, the next task at the top of the queue is recommended. The almost ready to publish. There is one small glitch for the emulator that makes the create button shift slightly to the left, but this is not on the internet emulator. Next I need to test on a real device.